requirements.txt
python-telegram-bot>=21.6
httpx>=0.27
python-dotenv>=1.0
pikepdf>=9.0
